[SC-755] PDSEvent returns a nonce of BigInteger.

+++ java-score/src/main/java/com/iconloop/score/pds/LabelInfo.java
@@ -91,8 +91,8 @@ public class LabelInfo {
         return (this.expireAt.isEmpty()) ? BigInteger.ZERO : new BigInteger(this.expireAt);
     }
 
-    public int getNonce() {
-        return this.nonce.intValue();
+    public BigInteger getNonce() {
+        return this.nonce;
     }
 
     public boolean checkNonce(BigInteger nonce) {
